+++ backend/fittoring/src/main/java/fittoring/config/S3Configuration.java
@@ -1,18 +1,31 @@
 package fittoring.config;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import software.amazon.awssdk.auth.credentials.DefaultCredentialsProvider;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 
+@RequiredArgsConstructor
 @Configuration
 public class S3Configuration {
+
+    private final S3Properties s3Properties;
 
     @Bean
     public S3Client s3Client() {
         return S3Client.builder()
                 .region(Region.AP_NORTHEAST_2)
+                .credentialsProvider(DefaultCredentialsProvider.create())
+                .build();
+    }
+
+    @Bean
+    public S3Presigner s3Presigner() {
+        return S3Presigner.builder()
+                .region(Region.of(s3Properties.getRegion()))
                 .credentialsProvider(DefaultCredentialsProvider.create())
                 .build();
     }
