--- backend/fittoring/src/main/java/fittoring/mentoring/infra/image/S3Config.java
@@ -1,29 +1,0 @@
-package fittoring.mentoring.infra.image;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.presigner.S3Presigner;
-
-@RequiredArgsConstructor
-@Configuration
-public class S3Config {
-
-    private final S3Properties s3Properties;
-
-    @Bean
-    public S3Client s3Client() {
-        return S3Client.builder()
-                .region(Region.of(s3Properties.getRegion()))
-                .build();
-    }
-
-    @Bean
-    public S3Presigner s3Presigner() {
-        return S3Presigner.builder()
-                .region(Region.of(s3Properties.getRegion()))
-                .build();
-    }
-}
